+++ tests/specs/bad-args.js
@@ -76,30 +76,6 @@
                     done();
                 });
             });
-            it('should catch when keyPath is an array and multiEntry=true', function (done) {
-                db.open({
-                    server: this.dbName,
-                    version: 2,
-                    schema: {
-                        test: {
-                            key: {
-                                keyPath: ['lastName', 'firstName']
-                            },
-                            indexes: {
-                                name: {
-                                    keyPath: ['lastName', 'firstName'],
-                                    multiEntry: true
-                                },
-                                lastName: {},
-                                firstName: {}
-                            }
-                        }
-                    }
-                }).catch(function (err) {
-                    expect(err.name).to.equal('InvalidAccessError');
-                    done();
-                });
-            });
         });
 
         describe('open: createSchema', function () {
@@ -124,7 +100,7 @@
                         }
                     }
                 }}).catch(function (err) {
-                    expect(err.name).to.equal('InvalidAccessError');
+                    expect(err.name).to.equal('InvalidAccessError'); // PhantomJS fails here, but this is correct per current and draft spec
                     done();
                 });
             });